[WIP] Add a kded module infra for fuse mount services
Summary:
This module watches the install directories of kded modules and curates a list of modules having plugin Category=MountService.
This list will be used by a kio kcm which will allow users to configure mount service for remote protocol.
Once loaded this module can be used in KIOSlave::listDir to mount the url and in KRun::resolveUrls to convert the
remote URL to local URL.

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22708